Refactor: 내 정보 관련 타입 파일 생성
