+++ app/_types/user.account.ts
@@ -1,0 +1,9 @@
+export interface ErrorResponse {
+  message: string;
+}
+
+export interface UserFormData {
+  newPassword: string;
+  nickname: string;
+  profileImageUrl?: string;
+}
